+++ client/src/components/Pages/Bursary/index.js
@@ -1,953 +1,393 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
+import React, { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
-import { Input, UploadFile, DatePicker, Select } from '../../Common/Inputs';
-import { Col, Row } from '../../Common/Grid';
-import * as S from './style';
+import axios from 'axios';
+import moment from 'moment';
+
 import * as T from '../../Common/Typography';
-import Button from '../../Common/ButtonNew';
-import {
-  API_MY_BURSARY,
-  API_SINGLE_BURSARY,
-} from '../../../constants/apiRoutes';
-import Notification from '../../Common/Notification';
-
+import { Col, Row } from '../../Common/Grid';
+import FAQ from '../../Common/FAQ';
+import BursaryApplication from './Application';
 import BursarySuccess from './Success';
+import {
+  API_BURSARY_APPLICATIONS_STATUS,
+  API_BURSARY_WINDOWS,
+  API_MY_PROFILE_URL,
+} from '../../../constants/apiRoutes';
+import { INTERN_SIGNUP_BURSARY } from '../../../constants/navRoutes';
+import Button from '../../Common/ButtonNew';
+import * as S from './style';
 
-import types from '../../../constants/types';
-import { BURSARY_SUCCESS } from '../../../constants/navRoutes';
+const Bursary = () => {
+  const [bursary, setBursary] = useState();
+  const [window, setWindow] = useState();
+  const [over14Days, setOver14Days] = useState();
 
-const { validate, internBursary } = require('../../../validation');
-
-const getCleanData = (d = {}) => ({
-  // bursary data
-  typeOfSchool: d.typeOfSchool || '',
-  typeOfSchoolOther: d.typeOfSchoolOther || '',
-  highestLevelOfQualifications: d.highestLevelOfQualifications || '',
-  highestLevelOfQualificationsOther: d.highestLevelOfQualificationsOther || '',
-  describeMainIncomeEarnerMainJob: d.describeMainIncomeEarnerMainJob || '',
-  numberOfPeopleKnowBefore16: d.numberOfPeopleKnowBefore16 || '',
-  typeOfUniversity: d.typeOfUniversity || '',
-  eligibleForFreeSchoolMeals: d.eligibleForFreeSchoolMeals || '',
-  comingFromLowerSociolEconomicBackground:
-    d.comingFromLowerSociolEconomicBackground || '',
-  householdMembersSpeakOtherLanguage:
-    d.householdMembersSpeakOtherLanguage || '',
-  householdMembersSpeakOtherLanguageYes:
-    d.householdMembersSpeakOtherLanguageYes || '',
-  annualHouseholdIncome: d.annualHouseholdIncome || '',
-  statusOfHome: d.statusOfHome || '',
-  statusOfHomeOther: d.statusOfHomeOther || '',
-  anyHouseholdReceive: d.anyHouseholdReceive || '',
-  benefitFromNepotism: d.benefitFromNepotism || '',
-  peopleYouKnowSocially: d.peopleYouKnowSocially || '',
-  accentAffectsPotentialEmployers: d.accentAffectsPotentialEmployers || '',
-  parentsSupportiveOfCareer: d.parentsSupportiveOfCareer || '',
-
-  // profile data
-  organisation: d.organisation || '',
-  internshipContact: d.internshipContact || {
-    name: '',
-    email: '',
-    phoneNumber: '',
-  },
-  internshipStartDate: d.internshipStartDate || null,
-  internshipEndDate: d.internshipEndDate || null,
-  internshipOfficeAddress: d.internshipOfficeAddress || {
-    addressline1: '',
-    addressline2: '',
-    city: '',
-    postcode: '',
-  },
-  offerLetter: d.offerLetter || {
-    fileName: '',
-  },
-});
-
-const Bursary = props => {
   const history = useHistory();
-
-  const [state, setState] = useState(getCleanData({}));
-
-  const [errors, setErrors] = useState({});
-  const [mainError, setMainError] = useState();
-  const [loading, setLoading] = useState(false);
-  const [notificationOpen, setNotificationOpen] = useState(false);
-  const [prevData, setPrevData] = useState({});
-  const [fetchingData, setFetchingData] = useState(true);
-
   useEffect(() => {
-    window.scrollTo({
-      left: 0,
-      top: 0,
-    });
-  }, []);
-
-  const _validate = async () => {
-    const { errors: _errors } = await validate({
-      schema: internBursary(prevData),
-      data: { ...state },
-    });
-
-    let e = _errors;
-
-    if (
-      (prevData.offerLetter && state.offerLetter.deleted) ||
-      !state.offerLetter
-    ) {
-      e = e
-        ? { ...e, offerLetter: 'Proof of internship file is required' }
-        : { offerLetter: 'Proof of internship file is required' };
-    }
-
-    return e;
-  };
-
-  const onInputChange = e => {
-    const { value, name } = e.target;
-
-    setErrors(_errors => ({ ..._errors, [name]: '' }));
-    return setState(_state => ({ ..._state, [name]: value }));
-  };
-
-  const uploadFile = async (file = {}) => {
-    try {
-      const generatedName = `${props.id}/${Date.now()}.${file.name}`;
-      const {
-        data: { signedUrl },
-      } = await axios.get(`/api/upload/signed-url?fileName=${generatedName}`);
-      const headers = {
-        'Content-Type': 'application/octet-stream',
-      };
-
-      await axios.put(signedUrl, file, {
-        headers,
-      });
-
-      return {
-        fileName: generatedName,
-        new: true,
-        uploaded: true,
-        preview: file.preview,
-      };
-    } catch (e) {
-      return setMainError(e.message);
-    }
-  };
-
-  const onSave = async () => {
-    try {
-      let _offerLetter;
-
-      setErrors({});
-
-      const _errors = await _validate();
-
-      setErrors(_errors || {});
-
-      if (_errors) {
-        setMainError('Must fill all required fields');
-        return;
-      }
-      setMainError();
-
-      setLoading(true);
-
-      if (
-        state.offerLetter &&
-        state.offerLetter.new &&
-        !state.offerLetter.uploaded
-      ) {
-        const promiseArr = [];
-
-        if (state.offerLetter && state.offerLetter.new) {
-          promiseArr.push(
-            uploadFile(state.offerLetter).catch(err =>
-              setErrors(e => ({ ...e, offerLetter: err.message })),
-            ),
-          );
-        } else {
-          promiseArr.push(Promise.resolve());
-        }
-
-        [_offerLetter] = await Promise.all(promiseArr);
-      }
-
-      await axios.patch(
-        `${API_SINGLE_BURSARY.replace(':id', prevData.id)}?profile=true`,
-        {
-          ...state,
-          offerLetter: _offerLetter || state.offerLetter,
-          prevOfferLetterToDelete:
-            state.offerLetter &&
-            state.offerLetter.new &&
-            prevData.offerLetter &&
-            prevData.offerLetter.fileName &&
-            prevData.offerLetter.fileName,
-        },
-      );
-      setFetchingData(true);
-      setNotificationOpen(true);
-    } catch (e) {
-      setMainError(e.response.data.error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    if (!state.typeOfSchool || !state.typeOfSchool.includes('Other')) {
-      setState(_state => ({ ..._state, typeOfSchoolOther: '' }));
-    }
-  }, [state.typeOfSchool]);
-
-  useEffect(() => {
-    if (
-      !state.householdMembersSpeakOtherLanguage ||
-      !state.householdMembersSpeakOtherLanguage.includes('Yes')
-    ) {
-      setState(_state => ({
-        ..._state,
-        householdMembersSpeakOtherLanguageYes: '',
-      }));
-    }
-  }, [state.householdMembersSpeakOtherLanguage]);
-
-  useEffect(() => {
-    if (
-      !state.highestLevelOfQualifications ||
-      !state.highestLevelOfQualifications.includes('Other')
-    ) {
-      setState(_state => ({
-        ..._state,
-        highestLevelOfQualificationsOther: '',
-      }));
-    }
-  }, [state.highestLevelOfQualifications]);
-
-  useEffect(() => {
-    if (!state.statusOfHome || !state.statusOfHome.includes('Other')) {
-      setState(_state => ({
-        ..._state,
-        statusOfHomeOther: '',
-      }));
-    }
-  }, [state.statusOfHome]);
-
-  useEffect(() => {
-    const getData = async () => {
-      const { data: bursary } = await axios.get(
-        `${API_MY_BURSARY}?profile=true`,
-      );
-      setFetchingData(false);
-
-      setState(getCleanData(bursary || {}));
-      setPrevData({
-        ...getCleanData(bursary || {}),
-        id: bursary && bursary._id,
-      });
+    const getBursaryApplication = async () => {
+      const { data } = await axios.get(API_BURSARY_APPLICATIONS_STATUS);
+      setBursary(data);
     };
 
-    getData();
+    const getWindows = async () => {
+      const { data } = await axios.get(API_BURSARY_WINDOWS);
+      if (data && data.length) {
+        const _window = data
+          .filter(({ startDate, endDate }) => {
+            if (
+              moment(startDate).valueOf() <= moment().valueOf() &&
+              moment(endDate).valueOf() >= moment().valueOf()
+            ) {
+              return true;
+            }
+            return false;
+          })
+          .sort(
+            (a, b) =>
+              moment(b.startDate).valueOf() - moment(a).valueOf(a.startDate),
+          );
+        setWindow(_window && _window[0]);
+      }
+    };
+
+    const getProfile = async () => {
+      const {
+        data: { profile },
+      } = await axios.get(API_MY_PROFILE_URL);
+
+      if (profile && profile.internshipEndDate && profile.internshipStartDate) {
+        const internshipDays = moment(profile.internshipEndDate)
+          .startOf()
+          .diff(moment(profile.internshipStartDate).startOf(), 'days');
+
+        setOver14Days(internshipDays >= 14);
+      }
+    };
+
+    getProfile();
+    getBursaryApplication();
+    getWindows();
   }, []);
 
-  const done = () => {
-    history.push(BURSARY_SUCCESS);
-  };
-
   return (
-    <div style={{ marginTop: '4rem', paddingBottom: '5rem' }}>
-      <Row>
-        <Col w={[4, 12, 6]}>
-          <T.H2 color="blue" mb={6}>
-            Bursary
-          </T.H2>
-          <T.H5 color="pink" mb={3}>
-            Please double check and update any relevant details below and then
-            click submit.
-          </T.H5>
-        </Col>
-      </Row>
+    <div>
+      <S.BursaryPageHeader>
+        <T.H2 color="blue">Bursary</T.H2>
 
-      <Row>
-        <Col w={[4, 8, 4]} mt={6}>
-          <Select
-            options={types.bursaryTypes.typeOfSchool.map(e => ({
-              label: e,
-              value: e,
-            }))}
-            label="What type of school did you mainly attend between the ages of 11 and 16?"
-            allowClear
-            onChange={value =>
-              setState(_state => ({ ..._state, typeOfSchool: value }))
-            }
-            value={state.typeOfSchool}
-            error={errors.typeOfSchool}
-          />
-
-          {state.typeOfSchool && state.typeOfSchool.includes('Other') && (
-            <Input
-              onChange={onInputChange}
-              value={state.typeOfSchoolOther}
-              label="Please specify"
-              name="typeOfSchoolOther"
-              error={errors.typeOfSchoolOther}
-            />
+        <div>
+          {bursary ? (
+            <S.BursaryButtonWrapperDesktop>
+              <Button
+                type="tertiary"
+                withGraphic
+                style={{ paddingLeft: '10px', paddingRight: '10px' }}
+              >
+                YOUR APPLICATION IS UNDER CONSIDERATION
+              </Button>
+            </S.BursaryButtonWrapperDesktop>
+          ) : (
+            <S.BursaryButtonWrapperDesktop>
+              {over14Days && (
+                <>
+                  <Button
+                    type="secondary"
+                    withGraphic
+                    onClick={() => history.push(INTERN_SIGNUP_BURSARY)}
+                  >
+                    APPLY FOR BURSARY
+                  </Button>
+                  {window ? (
+                    <T.PS color="gray3">
+                      Application deadline for the next round of bursaries is{' '}
+                      <T.PSBold>
+                        {moment(window.endDate)
+                          .endOf()
+                          .format('Do MMMM YYYY')}
+                        .
+                      </T.PSBold>
+                    </T.PS>
+                  ) : (
+                    <T.PS color="gray3">
+                      Sorry!, no open applications for now
+                    </T.PS>
+                  )}
+                </>
+              )}
+            </S.BursaryButtonWrapperDesktop>
           )}
+        </div>
+      </S.BursaryPageHeader>
+
+      {!bursary && !over14Days && (
+        <T.PS color="pink" mt={2}>
+          You are currently able to stay with PressPad hosts for up to 2 weeks
+          completely free so you do not need to apply for a bursary right now as
+          this covers the period of your internship
+        </T.PS>
+      )}
+
+      <T.H3C color="pink" mt={8} mb={3}>
+        How does it work?
+      </T.H3C>
+
+      <T.PXL bold color="blue" mb={3}>
+        The PressPad bursary is funded directly from our corporate donors,
+        grants and any revenue PressPad can generate by delivering training or
+        consultancy work. As per our commitment to social impact, we also donate
+        15% of all profits we receive from our hosting platform to our bursary
+        impact fund.
+      </T.PXL>
+      <T.P color="gray3" mb={7}>
+        Our bursary awards cover 50% or 100% of the cost of your placement, with
+        the maximum bursary in real money terms capped at £840. We ask
+        candidates questions to establish if they may be eligible for a bursary
+        award so we can identify those who lack socio-economic privilege and are
+        the most in need of our support, which includes our widening
+        participation* work. Candidates may apply for multiple awards in any
+        given year and PressPad will assess each application independently.
+      </T.P>
+
+      <T.H3C color="pink" mb={6}>
+        FAQS
+      </T.H3C>
+      <Row inner>
+        <Col w={[4, 12, 6]}>
+          <FAQ
+            title="How do we decide who gets a PressPad Bursary?"
+            content={
+              <>
+                <T.PBold color="blue">
+                  We have a points system in place so the allocation of
+                  bursaries is fair and unbiased. Each candidate will be awarded
+                  a score upon completion of their bursary application.
+                </T.PBold>
+                <Row inner mt={4}>
+                  <Col w={[4, 4, 4]}>
+                    <T.H6C color="pink">100% Bursary</T.H6C>
+                  </Col>
+                  <Col w={[4, 8, 8]}>
+                    <T.P color="gray3">
+                      700+ points required = accommodation costs covered up to
+                      the value of £840.
+                    </T.P>
+                  </Col>
+                </Row>
+                <Row inner mt={4}>
+                  <Col w={[4, 4, 4]}>
+                    <T.H6C color="pink">50% Bursary</T.H6C>
+                  </Col>
+                  <Col w={[4, 8, 8]}>
+                    <T.P color="gray3">
+                      500-700 points = accommodation costs covered up to the
+                      value of £420.
+                    </T.P>
+                  </Col>
+                </Row>
+              </>
+            }
+            colorChevron="lightBlue"
+          />
+        </Col>
+        <Col w={[4, 12, 6]}>
+          <FAQ
+            title="What if I don’t fulfill YOUR criteria but am still in need?"
+            content={
+              <T.P color="gray3">
+                When a candidate completes our bursary application and scores
+                less than 500 points they will not be eligible for a bursary
+                award. However, we will always take into account unique
+                circumstances that require further review and welcome candidates
+                to appeal should they feel a decision is unfair.
+              </T.P>
+            }
+            colorChevron="lightBlue"
+          />
+        </Col>
+
+        <Col w={[4, 12, 6]}>
+          <FAQ
+            title="If I get awarded a bursary, what happens next? Are there any extra obligations I have?"
+            content={
+              <T.P color="gray3">
+                We could ask you to write a blog about your experience, join our
+                alumni community and fill out our before and after surveys (and
+                preferably follow-up surveys) so we can assess the bursary’s
+                impact.
+              </T.P>
+            }
+          />
+        </Col>
+        <Col w={[4, 12, 6]}>
+          <FAQ
+            title="How much is awarded as part of a bursary?"
+            content={
+              <>
+                <Row inner mt={4}>
+                  <Col w={[4, 4, 4]}>
+                    <T.H6C color="pink">100% Bursary</T.H6C>
+                  </Col>
+                  <Col w={[4, 8, 8]}>
+                    <T.P color="gray3">
+                      700+ points required = accommodation costs covered up to
+                      the value of £840.
+                    </T.P>
+                  </Col>
+                </Row>
+                <Row inner mt={4}>
+                  <Col w={[4, 4, 4]}>
+                    <T.H6C color="pink">50% Bursary</T.H6C>
+                  </Col>
+                  <Col w={[4, 8, 8]}>
+                    <T.P color="gray3">
+                      500-700 points = accommodation costs covered up to the
+                      value of £420.
+                    </T.P>
+                  </Col>
+                </Row>
+              </>
+            }
+          />
+        </Col>
+
+        <Col w={[4, 12, 6]}>
+          <FAQ
+            title="Are there other ways of funding my hosting with PressPad?"
+            content={
+              <T.P color="gray3">
+                Media companies can also purchase credit to support their
+                interns by subsidising the cost of accommodation. Please contact
+                the company to find out if credit is available.{' '}
+              </T.P>
+            }
+            colorChevron="lightBlue"
+          />
+        </Col>
+
+        <Col w={[4, 12, 6]}>
+          <FAQ
+            title="How do we fund our bursaries?"
+            content={
+              <T.P color="gray3">
+                PressPad has a mission lock written into our governing document
+                that states that at least 10% of our income goes to the PressPad
+                bursary scheme so that everyone can access and use our service.
+              </T.P>
+            }
+            colorChevron="lightBlue"
+          />
+        </Col>
+
+        <Col w={[4, 12, 6]}>
+          <FAQ
+            title="Can I get a bursary from PressPad for travel and other living expenses?"
+            content={
+              <>
+                <T.P color="gray3">
+                  <T.PBold as="span" color="blue">
+                    London Bursary Allowance:
+                  </T.PBold>{' '}
+                  The PressPad London Bursary allowance provides an additional
+                  20% of the bursary award (capped at £168) if a candidate is
+                  living and working within Zones 1-4 of the London area.
+                  Candidates may be eligible for this if they have met the
+                  requirements of a 100% bursary award and will live and work in
+                  the London area for the duration of their internship.
+                </T.P>
+                <T.P color="gray3" mt={3}>
+                  <T.PBold as="span" color="blue">
+                    Travel Expenses:{' '}
+                  </T.PBold>
+                  We’re working hard to secure extra funding to help cover the
+                  cost of travel to and from your PressPad accommodation and
+                  placement, however, we are currently not able to provide
+                  financial support for these expenses right now. Always ask if
+                  your work experience or internship includes expenses or not.
+                  If in doubt check the government guidance
+                  <T.Link to="https://www.gov.uk/guidance/national-minimum-wage-work-experience-and-internships#arrangements-which-dont-qualify-for-the-minimum-wage" />
+                  Hint: The examples at the bottom of the page are the most
+                  useful we’ve found!
+                </T.P>
+              </>
+            }
+          />
+        </Col>
+        <Col w={[4, 12, 6]}>
+          <FAQ
+            title="*What does Widening Participation mean?"
+            content={
+              <>
+                <T.P color="blue" mt={4}>
+                  Diversity and Inclusion are firmly embedded into how PressPad
+                  is run. We want to help remove any barriers that
+                  underrepresented communities face when pursuing a career in
+                  journalism. To achieve this, we allocate at least 20% of our
+                  bursary impact fund towards widening participation.
+                </T.P>
+                <T.P color="blue" mt={4}>
+                  Widening participation is where we specifically reach out to
+                  groups who are from underrepresented backgrounds e.g. young
+                  carers, people with a disability or those with an ethnic
+                  background that have barriers to entering the media industry.
+                </T.P>
+                <T.P color="blue" mt={4}>
+                  Our primary objective is, and will always be, to help those
+                  from low socio-economic groups with the necessary financial
+                  support to pursue a career in the media. But we also
+                  understand that disadvantage, prejudice and identity are
+                  complex and intersectional - and for that reason we committed
+                  to grow our widening participation fund as the organisation
+                  grows and gains more support.
+                </T.P>
+                <T.P color="blue" mt={4}>
+                  We feel passionate about firmly embedding these principles
+                  into our outreach to help close the access gap so that
+                  everyone regardless of income, race, religion or gender can
+                  follow an exciting and enriching path in journalism.
+                </T.P>
+              </>
+            }
+          />
         </Col>
       </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.highestLevelOfQualifications.map(
-                e => ({
-                  label: e,
-                  value: e,
-                }),
+      {bursary ? (
+        <Button
+          type="tertiary"
+          withGraphic
+          style={{ paddingLeft: '10px', paddingRight: '10px' }}
+        >
+          YOUR APPLICATION IS UNDER CONSIDERATION
+        </Button>
+      ) : (
+        <S.BursaryButtonWrapperTablet>
+          {over14Days && (
+            <>
+              <Button
+                type="secondary"
+                withGraphic
+                onClick={() => history.push(INTERN_SIGNUP_BURSARY)}
+              >
+                APPLY FOR BURSARY
+              </Button>
+              {window ? (
+                <T.PS color="gray3">
+                  Application deadline for the next round of bursaries is{' '}
+                  <T.PSBold>
+                    {moment(window.endDate)
+                      .endOf()
+                      .format('Do MMMM YYYY')}
+                    .
+                  </T.PSBold>
+                </T.PS>
+              ) : (
+                <T.PS color="gray3">Sorry!, no open applications for now</T.PS>
               )}
-              label="What is the highest level of qualifications achieved by either of your parent(s) or guardian(s) by the time you were 18?"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  highestLevelOfQualifications: value,
-                }))
-              }
-              value={state.highestLevelOfQualifications}
-              error={errors.highestLevelOfQualifications}
-            />
-
-            {state.highestLevelOfQualifications &&
-              state.highestLevelOfQualifications.includes('Other') && (
-                <Input
-                  onChange={onInputChange}
-                  value={state.highestLevelOfQualificationsOther}
-                  label="Please specify"
-                  name="highestLevelOfQualificationsOther"
-                  error={errors.highestLevelOfQualificationsOther}
-                />
-              )}
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.describeMainIncomeEarnerMainJob.map(
-                e => ({
-                  label: e,
-                  value: e,
-                }),
-              )}
-              label="Thinking back to when you were aged about 14, which best describes the sort of work the main/ highest income earner in your household did in their main job?"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  describeMainIncomeEarnerMainJob: value,
-                }))
-              }
-              value={state.describeMainIncomeEarnerMainJob}
-              error={errors.describeMainIncomeEarnerMainJob}
-            />
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.numberOfPeopleKnowBefore16.map(e => ({
-                label: e,
-                value: e,
-              }))}
-              label="How many people did you know before the age of 16 that work in the industry you're pursuing a career in?"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  numberOfPeopleKnowBefore16: value,
-                }))
-              }
-              value={state.numberOfPeopleKnowBefore16}
-              error={errors.numberOfPeopleKnowBefore16}
-            />
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.typeOfUniversity.map(e => ({
-                label: e,
-                value: e,
-              }))}
-              label="If you attended university, what type of did you attend?"
-              helperText="(select as many as you like)"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  typeOfUniversity: value,
-                }))
-              }
-              multi
-              height="auto"
-              value={state.typeOfUniversity}
-              error={errors.typeOfUniversity}
-            />
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.eligibleForFreeSchoolMeals.map(e => ({
-                label: e,
-                value: e,
-              }))}
-              label="If you finished school after 1980, were you eligible for Free School Meals at any point during your school years? "
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  eligibleForFreeSchoolMeals: value,
-                }))
-              }
-              value={state.eligibleForFreeSchoolMeals}
-              error={errors.eligibleForFreeSchoolMeals}
-              extraInfo="Free School Meals are a statutory benefit available to school-aged children from families who receive other qualifying benefits and who have been through the relevant registration process. It does not include those who receive meals at school through other means (e.g. boarding school)."
-            />
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.comingFromLowerSociolEconomicBackground.map(
-                e => ({
-                  label: e,
-                  value: e,
-                }),
-              )}
-              label="Compared to people in general, would you describe yourself as coming from a lower socio-economic background?"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  comingFromLowerSociolEconomicBackground: value,
-                }))
-              }
-              value={state.comingFromLowerSociolEconomicBackground}
-              error={errors.comingFromLowerSociolEconomicBackground}
-            />
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.householdMembersSpeakOtherLanguage.map(
-                e => ({
-                  label: e,
-                  value: e,
-                }),
-              )}
-              label="Do household members speak and/or write a language other than English?"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  householdMembersSpeakOtherLanguage: value,
-                }))
-              }
-              value={state.householdMembersSpeakOtherLanguage}
-              error={errors.householdMembersSpeakOtherLanguage}
-            />
-            {state.householdMembersSpeakOtherLanguage &&
-              state.householdMembersSpeakOtherLanguage.includes('Yes') && (
-                <Input
-                  onChange={onInputChange}
-                  value={state.householdMembersSpeakOtherLanguageYes}
-                  label="Please specify"
-                  name="householdMembersSpeakOtherLanguageYes"
-                  error={errors.householdMembersSpeakOtherLanguageYes}
-                />
-              )}
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.annualHouseholdIncome.map(e => ({
-                label: e,
-                value: e,
-              }))}
-              label="What is your annual household income after tax?"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  annualHouseholdIncome: value,
-                }))
-              }
-              value={state.annualHouseholdIncome}
-              error={errors.annualHouseholdIncome}
-            />
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.statusOfHome.map(e => ({
-                label: e,
-                value: e,
-              }))}
-              label="What is the status of your home?"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  statusOfHome: value,
-                }))
-              }
-              value={state.statusOfHome}
-              error={errors.statusOfHome}
-            />
-            {state.statusOfHome && state.statusOfHome.includes('Other') && (
-              <Input
-                onChange={onInputChange}
-                value={state.statusOfHomeOther}
-                label="Please specify"
-                name="statusOfHomeOther"
-                error={errors.statusOfHomeOther}
-              />
-            )}
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.anyHouseholdReceive.map(e => ({
-                label: e,
-                value: e,
-              }))}
-              label="How many people in your household at present receive any of the following?"
-              helperText="(select as many as you like)"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  anyHouseholdReceive: value,
-                }))
-              }
-              height="auto"
-              multi
-              value={state.anyHouseholdReceive}
-              error={errors.anyHouseholdReceive}
-            />
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.benefitFromNepotism.map(e => ({
-                label: e,
-                value: e,
-              }))}
-              label="Have you ever benefited from nepotism in your work life?"
-              extraInfo="Nepotism is what happens when someone receives favouritism or opportunities because of a family relationship in the context of jobs, business and politics. We think it can also refer to benefits received in connection with a family friend. For example: If you get an internship at a news organisation because your Dad’s university friend works there, or you get a chance to have a coffee with a journalist because your cousin's colleague knows them. We believe nepotism refers to benefiting from any connection that you haven’t made for yourself but that was passed down to you."
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  benefitFromNepotism: value,
-                }))
-              }
-              value={state.benefitFromNepotism}
-              error={errors.benefitFromNepotism}
-            />
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.peopleYouKnowSocially.map(e => ({
-                label: e,
-                value: e,
-              }))}
-              label="Which of these people do you know socially?"
-              helperText="(Select all that apply)"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  peopleYouKnowSocially: value,
-                }))
-              }
-              height="auto"
-              multi
-              value={state.peopleYouKnowSocially}
-              error={errors.peopleYouKnowSocially}
-            />
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.accentAffectsPotentialEmployers.map(
-                e => ({
-                  label: e,
-                  value: e,
-                }),
-              )}
-              label="Do you feel your accent affects the way potential employers view you?"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  accentAffectsPotentialEmployers: value,
-                }))
-              }
-              value={state.accentAffectsPotentialEmployers}
-              error={errors.accentAffectsPotentialEmployers}
-            />
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-      <Row mb={6}>
-        <S.IllCareWrapper>
-          <Col w={[4, 8, 4]} mt={6}>
-            <Select
-              options={types.bursaryTypes.parentsSupportiveOfCareer.map(e => ({
-                label: e,
-                value: e,
-              }))}
-              label="Do you feel your parents are supportive of your chosen career?"
-              allowClear
-              onChange={value =>
-                setState(_state => ({
-                  ..._state,
-                  parentsSupportiveOfCareer: value,
-                }))
-              }
-              value={state.parentsSupportiveOfCareer}
-              error={errors.parentsSupportiveOfCareer}
-            />
-          </Col>
-        </S.IllCareWrapper>
-      </Row>
-
-      <Row>
-        <Col w={[4, 8, 8]}>
-          <T.H5 color="blue" mt={6}>
-            Placement Details
-          </T.H5>
-          <T.P color="gray3" mt={2}>
-            In order to apply for a bursary, you nee dot have an internship,
-            work placements or contract lined up. Please check that the below
-            internship details are still correct.
-          </T.P>
-        </Col>
-      </Row>
-
-      <Row>
-        <Col w={[4, 6, 6]} mt={6}>
-          <Input
-            onChange={onInputChange}
-            value={state.organisation}
-            label="Name of organisation"
-            name="organisation"
-            error={errors.organisation}
-          />
-        </Col>
-
-        <Col w={[4, 6, 6]} mt={6}>
-          <Input
-            onChange={e => {
-              e.persist();
-              setState(_state => ({
-                ..._state,
-                internshipContact: {
-                  ...state.internshipContact,
-                  name: e.target.value,
-                },
-              }));
-            }}
-            value={state.internshipContact && state.internshipContact.name}
-            label="Contact name"
-            placeholder="Contact name..."
-            name="name"
-            error={errors.internshipContact && errors.internshipContact.name}
-          />
-        </Col>
-      </Row>
-
-      <Row>
-        <Col w={[4, 6, 6]} mt={6}>
-          <Input
-            onChange={e => {
-              e.persist();
-              setState(_state => ({
-                ..._state,
-                internshipContact: {
-                  ...state.internshipContact,
-                  email: e.target.value,
-                },
-              }));
-            }}
-            value={state.internshipContact && state.internshipContact.email}
-            label="Contact email"
-            placeholder="Contact email..."
-            name="email"
-            error={errors.internshipContact && errors.internshipContact.email}
-          />
-        </Col>
-        <Col w={[4, 6, 6]} mt={6}>
-          <Input
-            onChange={e => {
-              e.persist();
-              setState(_state => ({
-                ..._state,
-                internshipContact: {
-                  ...state.internshipContact,
-                  phoneNumber: e.target.value,
-                },
-              }));
-            }}
-            value={
-              state.internshipContact && state.internshipContact.phoneNumber
-            }
-            label="Contact number"
-            placeholder="Contact number..."
-            name="phoneNumber"
-            error={
-              errors.internshipContact && errors.internshipContact.phoneNumber
-            }
-          />
-        </Col>
-      </Row>
-
-      <Row>
-        <Col w={[4, 6, 6]} mt={6}>
-          <DatePicker
-            onChange={momentDate =>
-              setState(_state => ({
-                ..._state,
-                internshipStartDate: momentDate,
-              }))
-            }
-            value={state.internshipStartDate}
-            label="Start date"
-            error={errors.internshipStartDate}
-          />
-        </Col>
-
-        <Col w={[4, 6, 6]} mt={6}>
-          <DatePicker
-            onChange={momentDate =>
-              setState(_state => ({
-                ..._state,
-                internshipEndDate: momentDate,
-              }))
-            }
-            value={state.internshipEndDate}
-            label="End date"
-            error={errors.internshipEndDate}
-          />
-        </Col>
-      </Row>
-
-      <Row>
-        <Col w={[4, 6, 6]} mt={6}>
-          <Input
-            onChange={e => {
-              e.persist();
-              setState(_state => ({
-                ..._state,
-                internshipOfficeAddress: {
-                  ...state.internshipOfficeAddress,
-                  addressline1: e.target.value,
-                },
-              }));
-            }}
-            value={
-              state.internshipOfficeAddress &&
-              state.internshipOfficeAddress.addressline1
-            }
-            label="Address Line 1"
-            name="addressline1"
-            error={
-              errors.internshipOfficeAddress &&
-              errors.internshipOfficeAddress.addressline1
-            }
-          />
-        </Col>
-
-        <Col w={[4, 6, 6]} mt={6}>
-          <Input
-            onChange={e => {
-              e.persist();
-              setState(_state => ({
-                ..._state,
-                internshipOfficeAddress: {
-                  ...state.internshipOfficeAddress,
-                  addressline2: e.target.value,
-                },
-              }));
-            }}
-            value={
-              state.internshipOfficeAddress &&
-              state.internshipOfficeAddress.addressline2
-            }
-            label="Address Line 2"
-            name="addressline2"
-            error={
-              errors.internshipOfficeAddress &&
-              errors.internshipOfficeAddress.addressline2
-            }
-          />
-        </Col>
-      </Row>
-
-      <Row>
-        <Col w={[4, 6, 6]} mt={6}>
-          <Input
-            onChange={e => {
-              e.persist();
-              setState(_state => ({
-                ..._state,
-                internshipOfficeAddress: {
-                  ...state.internshipOfficeAddress,
-                  city: e.target.value,
-                },
-              }));
-            }}
-            value={
-              state.internshipOfficeAddress &&
-              state.internshipOfficeAddress.city
-            }
-            label="City"
-            name="city"
-            error={
-              errors.internshipOfficeAddress &&
-              errors.internshipOfficeAddress.city
-            }
-          />
-        </Col>
-
-        <Col w={[4, 6, 6]} mt={6}>
-          <Input
-            onChange={e => {
-              e.persist();
-
-              setState(_state => ({
-                ..._state,
-                internshipOfficeAddress: {
-                  ...state.internshipOfficeAddress,
-                  postcode: e.target.value,
-                },
-              }));
-            }}
-            value={
-              state.internshipOfficeAddress &&
-              state.internshipOfficeAddress.postcode
-            }
-            label="Postcode"
-            name="postcode"
-            error={
-              errors.internshipOfficeAddress &&
-              errors.internshipOfficeAddress.postcode
-            }
-          />
-        </Col>
-      </Row>
-
-      <Row mt={6}>
-        <Col w={[4, 8, 8]}>
-          <T.H5 color="blue" mt={4}>
-            Proof of internship
-          </T.H5>
-          <T.P color="gray3">
-            e.g. an offer letter or something similar that can be used as proof
-            of your internship
-          </T.P>
-        </Col>
-      </Row>
-
-      <UploadFile
-        mainText="Upload file by dragging here"
-        secondaryText="file size max 2mb"
-        type="file"
-        userId={props.id}
-        files={[state.offerLetter]}
-        setFiles={([offerLetter]) =>
-          setState(_state => ({
-            ..._state,
-            offerLetter,
-          }))
-        }
-        error={errors.offerLetter}
-      />
-
-      <Row mt={6}>
-        <Col w={[6, 12, 12]}>
-          {mainError && <T.PXS color="pink">{mainError}</T.PXS>}
-        </Col>
-      </Row>
-      <Row>
-        <Col w={[4, 8, 4]} mb={6} mbT={3}>
-          <Button
-            type="secondary"
-            onClick={onSave}
-            loading={loading}
-            disabled={loading || fetchingData}
-          >
-            Submit
-          </Button>
-        </Col>
-      </Row>
-
-      <Notification
-        open={notificationOpen}
-        setOpen={setNotificationOpen}
-        content="Changes saved"
-        cb={done}
-      />
+            </>
+          )}
+        </S.BursaryButtonWrapperTablet>
+      )}
     </div>
   );
 };
 
-export { BursarySuccess };
+export { BursaryApplication, BursarySuccess };
 export default Bursary;
