+++ client/src/constants/apiRoutes.js
@@ -200,6 +200,9 @@ const API_SINGLE_BURSARY = `${API_PREFIX}${SINGLE_BURSARY}`;
 const TOP_ADMIN_STATS = '/admin/summary';
 const API_TOP_ADMIN_STATS = `${API_PREFIX}${TOP_ADMIN_STATS}`;
 
+const BURSARY_APPLICATIONS_STATUS = '/bursary/applications/status';
+const API_BURSARY_APPLICATIONS_STATUS = `${API_PREFIX}${BURSARY_APPLICATIONS_STATUS}`;
+
 module.exports = {
   // CLIENT
   API_LOGIN_URL,
@@ -270,6 +273,7 @@ module.exports = {
   API_MY_BURSARY,
   API_SINGLE_BURSARY,
   API_TOP_ADMIN_STATS,
+  API_BURSARY_APPLICATIONS_STATUS,
 
   // SERVER
   LOGIN_URL,
@@ -340,4 +344,5 @@ module.exports = {
   MY_BURSARY,
   SINGLE_BURSARY,
   TOP_ADMIN_STATS,
+  BURSARY_APPLICATIONS_STATUS,
 };
