+++ server/database/queries/bursary/getBursary.js
@@ -173,6 +173,18 @@ const getBursaryApplicationInfo = id => {
   ]);
 };
 
+const getPendingBursaryApplicationsByUserId = id => {
+  return BursaryApplication.findOne({
+    intern: id,
+    status: {
+      $in: [
+        bursaryApplicationStatuses.request,
+        bursaryApplicationStatuses.preApproved,
+      ],
+    },
+  });
+};
+
 const getBursaryByUserId = id => Bursary.findOne({ intern: id }).lean();
 
 module.exports = {
@@ -180,4 +192,5 @@ module.exports = {
   getBursaryApplications,
   getBursaryApplicationInfo,
   getBursaryByUserId,
+  getPendingBursaryApplicationsByUserId,
 };
