+++ server/router/index.js
@@ -76,6 +76,7 @@ const {
   upsertBursaryWindows,
   updateBursaryApplication,
   getBursaryApplicationInfo,
+  getMyBursaryApplicationStatus,
 } = require('../controllers/bursary');
 
 // IMPORT MIDDLEWARES
@@ -158,6 +159,7 @@ const {
   MY_BURSARY,
   SINGLE_BURSARY,
   TOP_ADMIN_STATS,
+  BURSARY_APPLICATIONS_STATUS,
 } = require('../../client/src/constants/apiRoutes');
 
 // add validation middleware
@@ -374,9 +376,15 @@ router.post(RESET_PASSWORD, resetPassword);
 router.post(SET_PASSWORD, setPassword);
 
 router.get(
+  BURSARY_APPLICATIONS_STATUS,
+  authentication,
+  getMyBursaryApplicationStatus,
+);
+
+router.get(
   BURSARY_WINDOWS,
   authentication,
-  authorization(['admin']),
+  authorization(['admin', 'intern']),
   getBursaryWindows,
 );
 router.put(
